+++ pycbc/neutron_stars/__init__.py
@@ -1,6 +1,6 @@
 import os.path
 
-from pycbc.neutron_stars.neutron_stars import *
+from pycbc.neutron_stars.ns_functions import *
 
 # Setup the directory with the NS equilibrium sequence(s)
 NS_SEQUENCE_FILE_DIRECTORY = os.path.join(os.path.dirname(__file__), 'ns_sequences')
