+++ setup.py
@@ -304,7 +304,8 @@ setup (
     packages = find_packages(),
     package_data = {'pycbc.workflow': find_files('pycbc/workflow'),
                     'pycbc.results': find_files('pycbc/results'),
-                    'pycbc.tmpltbank': find_files('pycbc/tmpltbank')},
+                    'pycbc.tmpltbank': find_files('pycbc/tmpltbank'),
+                    'pycbc.neutron_stars': find_files('pycbc/neutron_stars')},
     ext_modules = ext,
     python_requires='>=3.7',
     classifiers=[
